getAllPhysiciansController and Handler modified, gerAllPatientsHandler modified. Pagination folder created. paginationUsersHandler created.
